🚚 Move: 잘못 이동 시킨 CardDetail 이동
