+++ src/pages/Dashboard/components/SidePage/CardDetail/CardDetailStyled.ts
